+++ components/landing-page/contact-us.tsx
@@ -1,10 +1,10 @@
-import ContactForm from "./_components/contact-form";
+import ContactForm from "../forms/contact-form";
 
 export default function ContactUs() {
   return (
     <section
-      id='contact-us'
-      className='px-5 py-8 mx-auto md:px-8 md:py-12 max-w-7xl mx-auto'
+      id="contact-us"
+      className="px-5 py-8 md:px-8 md:py-12 max-w-7xl mx-auto"
     >
       <ContactForm />
     </section>
